refactor: Move wrapping helper fns to definitions/wrapping.ts
