+++ src/definitions/wrapping.ts
@@ -9,13 +9,21 @@ import { NexusExtendInputTypeDef } from './extendInputType'
 import { NexusExtendTypeDef } from './extendType'
 import { NexusInputObjectTypeDef } from './inputObjectType'
 import { NexusInterfaceTypeDef } from './interfaceType'
-import { NexusListDef } from './list'
-import { NexusNonNullDef } from './nonNull'
+import { NexusListDef, list } from './list'
+import { NexusNonNullDef, nonNull } from './nonNull'
 import { NexusNullDef } from './nullable'
 import { NexusObjectTypeDef } from './objectType'
 import { NexusScalarTypeDef } from './scalarType'
 import { NexusUnionTypeDef } from './unionType'
 import { NexusTypes, NexusWrappedSymbol } from './_types'
+import {
+  GraphQLNamedType,
+  GraphQLOutputType,
+  GraphQLInputType,
+  GraphQLType,
+  GraphQLNonNull,
+  GraphQLList,
+} from 'graphql'
 
 export type AllNexusNamedInputTypeDefs<T extends string = any> =
   | NexusInputObjectTypeDef<T>
@@ -154,4 +162,129 @@ export function isNexusPrintedGenTypingImport(obj: any): obj is PrintedGenTyping
 
 export function isNexusPlugin(obj: any): obj is NexusPlugin {
   return isNexusStruct(obj) && obj[NexusWrappedSymbol] === NexusTypes.Plugin
+}
+
+type NexusWrapKind = { type: 'List'; nullable: boolean } | { type: 'NamedType'; nullable: boolean }
+
+export function unwrapNexusDef(
+  typeDef: AllNexusTypeDefs | AllNexusArgsDefs | string,
+  nonNullDefault: boolean
+): { namedType: AllNexusNamedTypeDefs | AllNexusArgsDefs | string; wrapping: NexusWrapKind[] } {
+  const wrapping: NexusWrapKind[] = []
+  let namedType = typeDef
+
+  const isNamedType = (obj: any): obj is AllNexusNamedTypeDefs | NexusArgDef<any> | string =>
+    isNexusNamedTypeDef(obj) || isNexusArgDef(obj) || typeof obj === 'string'
+
+  if (isNamedType(typeDef)) {
+    return { namedType: typeDef, wrapping: [{ type: 'NamedType', nullable: !nonNullDefault }] }
+  }
+
+  while (isNexusWrappingType(namedType)) {
+    // nullable(list(Type))
+    if (isNexusNullTypeDef(namedType) && isNexusListTypeDef(namedType.ofType)) {
+      wrapping.unshift({ type: 'List', nullable: true })
+      namedType = namedType.ofType.ofType
+    }
+
+    // nullable(Type)
+    if (isNexusNullTypeDef(namedType) && isNamedType(namedType.ofType)) {
+      wrapping.unshift({ type: 'NamedType', nullable: true })
+      namedType = namedType.ofType
+      break
+    }
+
+    // nonNull(list(Type))
+    if (isNexusNonNullTypeDef(namedType) && isNexusListTypeDef(namedType.ofType)) {
+      wrapping.unshift({ type: 'List', nullable: false })
+      namedType = namedType.ofType.ofType
+    }
+
+    // nonNull(Type)
+    if (isNexusNonNullTypeDef(namedType) && isNamedType(namedType.ofType)) {
+      wrapping.unshift({ type: 'NamedType', nullable: false })
+      namedType = namedType.ofType
+      break
+    }
+
+    // list(...)
+    if (isNexusListTypeDef(namedType)) {
+      wrapping.unshift({ type: 'List', nullable: !nonNullDefault })
+      namedType = namedType.ofType
+    }
+
+    // Type
+    if (!isNexusWrappingType(namedType)) {
+      wrapping.unshift({ type: 'NamedType', nullable: !nonNullDefault })
+    }
+  }
+
+  return { namedType, wrapping }
+}
+
+/**
+ * Take a Nexus Wrapped Def, unwraps it, and rewraps it as a GraphQL wrapped type
+ * The outputted GraphQL type also reflects the nullability defaults
+ */
+export function rewrapAsGraphQLType(
+  nexusDef: AllNexusOutputTypeDefs | string,
+  baseType: GraphQLNamedType,
+  nonNullDefault: boolean
+): GraphQLOutputType
+export function rewrapAsGraphQLType(
+  nexusDef: AllNexusInputTypeDefs | string,
+  baseType: GraphQLNamedType,
+  nonNullDefault: boolean
+): GraphQLInputType
+export function rewrapAsGraphQLType(
+  nexusDef: AllNexusTypeDefs | string,
+  baseType: GraphQLNamedType,
+  nonNullDefault: boolean
+): GraphQLOutputType | GraphQLInputType {
+  const { wrapping } = unwrapNexusDef(nexusDef, nonNullDefault)
+
+  return wrapping.reduce<GraphQLType>((type, kind) => {
+    if (kind.type === 'NamedType' && kind.nullable === false) {
+      return GraphQLNonNull(type)
+    }
+
+    if (kind.type === 'List' && kind.nullable === true) {
+      return GraphQLList(type)
+    }
+
+    if (kind.type === 'List' && kind.nullable === false) {
+      return GraphQLNonNull(GraphQLList(type))
+    }
+
+    return type
+  }, baseType)
+}
+
+export function wrapAsNexusType(
+  baseType: AllNexusNamedTypeDefs | string,
+  wrapping: NexusWrapKind[]
+): AllNexusOutputTypeDefs
+export function wrapAsNexusType(
+  baseType: AllNexusNamedTypeDefs | string,
+  wrapping: NexusWrapKind[]
+): AllNexusInputTypeDefs
+export function wrapAsNexusType(
+  baseType: AllNexusNamedTypeDefs | string,
+  wrapping: NexusWrapKind[]
+): AllNexusTypeDefs {
+  return wrapping.reduce<any>((type, kind) => {
+    if (kind.type === 'NamedType' && kind.nullable === false) {
+      return nonNull(type)
+    }
+
+    if (kind.type === 'List' && kind.nullable === true) {
+      return list(type)
+    }
+
+    if (kind.type === 'List' && kind.nullable === false) {
+      return nonNull(list(type))
+    }
+
+    return type
+  }, baseType)
 }
